// server/interfaces/PrisonerAdvanceResponse.ts
export interface PrisonerAdvanceResponse {
  id: string
  prisonNumber: string
  legacyAdvanceNumber: number
  createdAt: string
  createdBy: string
  date: string
  advanceAmount: number
  paymentAmount: number
  startPayments: string
  reference: string
  advanceLocation: string
  status: string
}